+++ src/main/java/com/example/srvteam/dto/response/PontuacaoResponse.java
@@ -1,10 +1,13 @@
+
 package com.example.srvteam.dto.response;
+
+import java.math.BigDecimal;
 
 public class PontuacaoResponse {
     private Integer cdPontuacao;
     private Integer cdCompeticao;
     private Integer posicao;
-    private Integer pontuacao;
+        private BigDecimal pontuacao;
 
     public Integer getCdPontuacao() { return cdPontuacao; }
     public void setCdPontuacao(Integer cdPontuacao) { this.cdPontuacao = cdPontuacao; }
@@ -12,6 +15,6 @@ public class PontuacaoResponse {
     public void setCdCompeticao(Integer cdCompeticao) { this.cdCompeticao = cdCompeticao; }
     public Integer getPosicao() { return posicao; }
     public void setPosicao(Integer posicao) { this.posicao = posicao; }
-    public Integer getPontuacao() { return pontuacao; }
-    public void setPontuacao(Integer pontuacao) { this.pontuacao = pontuacao; }
+        public BigDecimal getPontuacao() { return pontuacao; }
+        public void setPontuacao(BigDecimal pontuacao) { this.pontuacao = pontuacao; }
 }
